+++ src/RayTracer/Transformation/IRotation.hpp
@@ -10,9 +10,9 @@
 namespace RayTracer {
     class IRotation {
         public:
-            virtual void rotate(float x, float y, float z) = 0;
-            virtual void rotateX(float x) = 0;
-            virtual void rotateY(float y) = 0;
-            virtual void rotateZ(float z) = 0;
+            virtual void rotate(double x, double y, double z) = 0;
+            virtual void rotateX(double x) = 0;
+            virtual void rotateY(double y) = 0;
+            virtual void rotateZ(double z) = 0;
     };
 };
